Create Op class (Short for Operations)
Provides a class to store all vector operations in as well as other functions not provided my the general Math class. It may be wise to import some sort of linear algebra class later and remove this file, as I doubt that modern developers are writing basic vector operations from scratch.